@@ -20,28 +20,28 @@ abstract class GeoCommands<K, V> {
   /// Returns the number of elements added to the sorted set.
   ///
   /// See https://redis.io/commands/geoadd
-  Future<int> geoadd(K key,
-      {GeoItem<V> item, Iterable<GeoItem<V>> items = const []});
+  Future<int?> geoadd(K key,
+      {GeoItem<V>? item, Iterable<GeoItem<V>> items = const []});
 
   /// Returns the distance between two members in the geospatial index
   /// represented by the sorted set stored at [key].
   ///
   /// See https://redis.io/commands/geodist
-  Future<double> geodist(K key, V member1, V member2, {GeoUnit unit});
+  Future<double?> geodist(K key, V member1, V member2, {GeoUnit? unit});
 
   /// Returns members of a geospatial index represented by the sorted set
   /// stored at [key] as standard geohash strings.
   ///
   /// See https://redis.io/commands/geohash
-  Future<List<String>> geohash(K key,
-      {V member, Iterable<V> members = const []});
+  Future<List<String?>?> geohash(K key,
+      {V? member, Iterable<V> members = const []});
 
   /// Returns the positions (longitude, latitude) of all the specified members
   /// of the geospatial index represented by the sorted set stored at [key].
   ///
   /// See https://redis.io/commands/geopos
-  Future<List<GeoPosition>> geopos(K key,
-      {V member, Iterable<V> members = const []});
+  Future<List<GeoPosition?>?> geopos(K key,
+      {V? member, Iterable<V> members = const []});
 
   /// Queries a geospatial index represented by the sorted set stored at [key]
   /// to fetch members matching a given maximum distance from a point.
@@ -49,13 +49,13 @@ abstract class GeoCommands<K, V> {
   /// See [georadiusStore].
   ///
   /// See https://redis.io/commands/georadius
-  Future<List<GeoradiusResult<V>>> georadius(
+  Future<List<GeoradiusResult<V>>?> georadius(
       K key, double longitude, double latitude, double radius, GeoUnit unit,
       {bool withCoord = false,
       bool withDist = false,
       bool withHash = false,
-      int count,
-      GeoOrder order});
+      int? count,
+      GeoOrder? order});
 
   /// Queries a geospatial index represented by the sorted set stored at [key]
   /// to fetch members matching a given maximum distance from a point and
@@ -64,9 +64,9 @@ abstract class GeoCommands<K, V> {
   /// See [georadius].
   ///
   /// See https://redis.io/commands/georadius
-  Future<int> georadiusStore(
+  Future<int?> georadiusStore(
       K key, double longitude, double latitude, double radius, GeoUnit unit,
-      {int count, GeoOrder order, K storeKey, K storeDistKey});
+      {int? count, GeoOrder? order, K? storeKey, K? storeDistKey});
 
   /// Queries a geospatial index represented by the sorted set stored at [key]
   /// to fetch members matching a given maximum distance from a member.
@@ -74,13 +74,13 @@ abstract class GeoCommands<K, V> {
   /// See [georadiusbymemberStore].
   ///
   /// See https://redis.io/commands/georadiusbymember
-  Future<List<GeoradiusResult<V>>> georadiusbymember(
+  Future<List<GeoradiusResult<V>>?> georadiusbymember(
       K key, V member, double radius, GeoUnit unit,
       {bool withCoord = false,
       bool withDist = false,
       bool withHash = false,
-      int count,
-      GeoOrder order});
+      int? count,
+      GeoOrder? order});
 
   /// Queries a geospatial index represented by the sorted set stored at [key]
   /// to fetch members matching a given maximum distance from a member and
@@ -89,9 +89,9 @@ abstract class GeoCommands<K, V> {
   /// See [georadiusbymember].
   ///
   /// See https://redis.io/commands/georadiusbymember
-  Future<int> georadiusbymemberStore(
+  Future<int?> georadiusbymemberStore(
       K key, V member, double radius, GeoUnit unit,
-      {int count, GeoOrder order, K storeKey, K storeDistKey});
+      {int? count, GeoOrder? order, K? storeKey, K? storeDistKey});
 }
 
 /// Metric units.
@@ -171,13 +171,13 @@ class GeoradiusResult<V> {
   final V member;
 
   /// The distance.
-  final double distance;
+  final double? distance;
 
   /// The hash.
-  final int hash;
+  final int? hash;
 
   /// The position.
-  final GeoPosition position;
+  final GeoPosition? position;
 
   /// Creates a [GeoradiusResult] instance.
   const GeoradiusResult(this.member, {this.distance, this.hash, this.position});
@@ -188,24 +188,20 @@ class GeoradiusResult<V> {
 }
 
 /// A mapper to be used with the GEOPOS command.
-class GeoPositionMapper implements Mapper<List<GeoPosition>> {
+class GeoPositionMapper implements Mapper<List<GeoPosition?>> {
   /// Creates a [GeoPositionMapper] instance.
   const GeoPositionMapper();
 
   @override
-  List<GeoPosition> map(covariant ArrayReply reply, RedisCodec codec) =>
-      reply.array
-          // ignore: avoid_as
-          .map((value) => _mapPosition(value as ArrayReply, codec))
-          .toList();
+  List<GeoPosition?> map(covariant ArrayReply reply, RedisCodec codec) => reply
+      .array
+      .map((value) =>
+          value is NullReply ? null : _mapPosition(value as ArrayReply, codec))
+      .toList();
 
   /// Maps a [reply] to `null` or a [GeoPosition] instance.
   GeoPosition _mapPosition(ArrayReply reply, RedisCodec codec) {
     final array = reply.array;
-
-    if (array == null) {
-      return null;
-    }
 
     final longitude = codec.decode<double>(array[0]);
     final latitude = codec.decode<double>(array[1]);
@@ -235,7 +231,6 @@ class GeoRadiusMapper<V> implements Mapper<List<GeoradiusResult<V>>> {
 
     for (final reply in reply.array) {
       if (withCoord || withDist || withHash) {
-        // ignore: avoid_as
         final result = _mapResult(reply as ArrayReply, codec);
         results.add(result);
       } else {
@@ -254,23 +249,22 @@ class GeoRadiusMapper<V> implements Mapper<List<GeoradiusResult<V>>> {
 
     final member = codec.decode<V>(array[index++]);
 
-    double distance;
+    double? distance;
     if (withDist) {
       distance = codec.decode<double>(array[index++]);
     }
 
-    int hash;
+    int? hash;
     if (withHash) {
       hash = codec.decode<int>(array[index++]);
     }
 
-    GeoPosition position;
+    GeoPosition? position;
     if (withCoord) {
-      // ignore: avoid_as
       position = _mapPosition(array[index++] as ArrayReply, codec);
     }
 
-    return GeoradiusResult<V>(member,
+    return GeoradiusResult<V>(member!,
         distance: distance, hash: hash, position: position);
   }
 
