@@ -3,14 +3,15 @@
 
 import 'package:test/test.dart';
 
+// ignore: directives_ordering
 import 'package:dartis/dartis.dart';
 
 import '../util.dart' show uuid;
 
 void main() {
-  PubSub<String, String> pubsub;
-  Client client;
-  Commands<String, String> commands;
+  late PubSub<String, String> pubsub;
+  late Client client;
+  late Commands<String, String> commands;
 
   setUp(() async {
     pubsub = await PubSub.connect<String, String>('redis://localhost:6379');
@@ -213,7 +214,9 @@ void main() {
 
       // Ping.
       final message = uuid();
-      pubsub..ping()..ping(message);
+      pubsub
+        ..ping()
+        ..ping(message);
 
       expect(
           pubsub.stream,
@@ -246,22 +249,27 @@ void main() {
 
     group('support', () {
       group('SubscriptionEvent', () {
-        test('toString', () {
-          const value = SubscriptionEvent<String>(null, null, null);
+        test('Non null toString', () {
+          const value = SubscriptionEvent<String>(null, 'channel', null);
           expect(value.toString(), startsWith('SubscriptionEvent<String>:'));
+        });
+        test('toString', () {
+          const value = SubscriptionEvent<String?>(null, null, null);
+          expect(value.toString(), startsWith('SubscriptionEvent<String?>:'));
         });
       });
       group('MessageEvent', () {
         test('toString', () {
-          const value = MessageEvent<String, String>(null, null);
-          expect(value.toString(), startsWith('MessageEvent<String, String>:'));
+          const value = MessageEvent<String?, String?>(null, null);
+          expect(
+              value.toString(), startsWith('MessageEvent<String?, String?>:'));
         });
       });
 
       group('PongEvent', () {
         test('toString', () {
-          const value = PongEvent<String>(null);
-          expect(value.toString(), startsWith('PongEvent<String>:'));
+          const value = PongEvent<String?>(null);
+          expect(value.toString(), startsWith('PongEvent<String?>:'));
         });
       });
     });
