@@ -3,13 +3,14 @@
 
 import 'package:test/test.dart';
 
+// ignore: directives_ordering
 import 'package:dartis/dartis.dart';
 
 import '../util.dart' show uuid;
 
 void main() {
-  Client client;
-  GeoCommands<String, String> commands;
+  late Client client;
+  late GeoCommands<String, String> commands;
 
   // Some items for testing.
   const item1 = GeoItem<String>(GeoPosition(-3.6827461, 40.4893538), 'Madrid');
@@ -111,15 +112,15 @@ void main() {
 
       // Get one position.
       var results = await commands.geopos(key1, member: 'Madrid');
-      expect(results[0].longitude, closeTo(-3.6827461, delta));
-      expect(results[0].latitude, closeTo(40.4893538, delta));
+      expect(results![0]!.longitude, closeTo(-3.6827461, delta));
+      expect(results[0]!.latitude, closeTo(40.4893538, delta));
 
       // Get some positions.
       results = await commands.geopos(key1, members: ['Valencia', 'Málaga']);
-      expect(results[0].longitude, closeTo(-0.3545661, delta));
-      expect(results[0].latitude, closeTo(39.4561165, delta));
-      expect(results[1].longitude, closeTo(-4.3971722, delta));
-      expect(results[1].latitude, closeTo(36.7585406, delta));
+      expect(results![0]!.longitude, closeTo(-0.3545661, delta));
+      expect(results[0]!.latitude, closeTo(39.4561165, delta));
+      expect(results[1]!.longitude, closeTo(-4.3971722, delta));
+      expect(results[1]!.latitude, closeTo(36.7585406, delta));
 
       // Get the position of a non existing member.
       final key2 = uuid();
@@ -128,7 +129,7 @@ void main() {
 
       results =
           await commands.geopos(key1, members: ['Valencia', key2, 'Málaga']);
-      expect(results[0], isNotNull);
+      expect(results![0], isNotNull);
       expect(results[1], isNull);
       expect(results[2], isNotNull);
 
@@ -149,7 +150,7 @@ void main() {
 
       results = await commands.georadius(
           key1, -0.2638491, 37.8234928, 200.0, GeoUnit.kilometer);
-      expect(results[0].member, 'Valencia');
+      expect(results![0].member, 'Valencia');
       expect(results[0].position, isNull);
       expect(results[0].hash, isNull);
       expect(results[0].distance, isNull);
@@ -158,9 +159,9 @@ void main() {
       results = await commands.georadius(
           key1, -0.2638491, 37.8234928, 200.0, GeoUnit.kilometer,
           withCoord: true);
-      expect(results[0].member, 'Valencia');
-      expect(results[0].position.longitude, closeTo(-0.3545661, delta));
-      expect(results[0].position.latitude, closeTo(39.4561165, delta));
+      expect(results![0].member, 'Valencia');
+      expect(results[0].position!.longitude, closeTo(-0.3545661, delta));
+      expect(results[0].position!.latitude, closeTo(39.4561165, delta));
       expect(results[0].hash, isNull);
       expect(results[0].distance, isNull);
 
@@ -168,9 +169,9 @@ void main() {
       results = await commands.georadius(
           key1, -0.2638491, 37.8234928, 200.0, GeoUnit.kilometer,
           withCoord: true, withHash: true);
-      expect(results[0].member, 'Valencia');
-      expect(results[0].position.longitude, closeTo(-0.3545661, delta));
-      expect(results[0].position.latitude, closeTo(39.4561165, delta));
+      expect(results![0].member, 'Valencia');
+      expect(results[0].position!.longitude, closeTo(-0.3545661, delta));
+      expect(results[0].position!.latitude, closeTo(39.4561165, delta));
       expect(results[0].hash, equals(1969199198234197));
       expect(results[0].distance, isNull);
 
@@ -178,9 +179,9 @@ void main() {
       results = await commands.georadius(
           key1, -0.2638491, 37.8234928, 200.0, GeoUnit.kilometer,
           withCoord: true, withHash: true, withDist: true);
-      expect(results[0].member, 'Valencia');
-      expect(results[0].position.longitude, closeTo(-0.3545661, delta));
-      expect(results[0].position.latitude, closeTo(39.4561165, delta));
+      expect(results![0].member, 'Valencia');
+      expect(results[0].position!.longitude, closeTo(-0.3545661, delta));
+      expect(results[0].position!.latitude, closeTo(39.4561165, delta));
       expect(results[0].hash, equals(1969199198234197));
       expect(results[0].distance, closeTo(181.7617, delta));
 
@@ -199,14 +200,14 @@ void main() {
       results = await commands.georadius(
           key1, -0.2638491, 37.8234928, 1000.0, GeoUnit.kilometer,
           order: GeoOrder.ascending);
-      expect(results[0].member, 'Valencia');
+      expect(results![0].member, 'Valencia');
       expect(results[1].member, 'Málaga');
       expect(results[2].member, 'Madrid');
 
       results = await commands.georadius(
           key1, -0.2638491, 37.8234928, 1000.0, GeoUnit.kilometer,
           order: GeoOrder.descending);
-      expect(results[0].member, 'Madrid');
+      expect(results![0].member, 'Madrid');
       expect(results[1].member, 'Málaga');
       expect(results[2].member, 'Valencia');
 
@@ -327,14 +328,14 @@ void main() {
       // Compute.
       var results = await commands.georadiusbymember(
           key1, 'Madrid', 0.0, GeoUnit.kilometer);
-      expect(results[0].member, 'Madrid');
+      expect(results![0].member, 'Madrid');
       expect(results[0].position, isNull);
       expect(results[0].hash, isNull);
       expect(results[0].distance, isNull);
 
       results = await commands.georadiusbymember(
           key1, 'Madrid', 400.0, GeoUnit.kilometer);
-      expect(results[0].member, 'Valencia');
+      expect(results![0].member, 'Valencia');
       expect(results[0].position, isNull);
       expect(results[0].hash, isNull);
       expect(results[0].distance, isNull);
@@ -343,9 +344,9 @@ void main() {
       results = await commands.georadiusbymember(
           key1, 'Madrid', 400.0, GeoUnit.kilometer,
           withCoord: true);
-      expect(results[0].member, 'Valencia');
-      expect(results[0].position.longitude, closeTo(-0.3545661, delta));
-      expect(results[0].position.latitude, closeTo(39.4561165, delta));
+      expect(results![0].member, 'Valencia');
+      expect(results[0].position!.longitude, closeTo(-0.3545661, delta));
+      expect(results[0].position!.latitude, closeTo(39.4561165, delta));
       expect(results[0].hash, isNull);
       expect(results[0].distance, isNull);
 
@@ -353,9 +354,9 @@ void main() {
       results = await commands.georadiusbymember(
           key1, 'Madrid', 400.0, GeoUnit.kilometer,
           withCoord: true, withHash: true);
-      expect(results[0].member, 'Valencia');
-      expect(results[0].position.longitude, closeTo(-0.3545661, delta));
-      expect(results[0].position.latitude, closeTo(39.4561165, delta));
+      expect(results![0].member, 'Valencia');
+      expect(results[0].position!.longitude, closeTo(-0.3545661, delta));
+      expect(results[0].position!.latitude, closeTo(39.4561165, delta));
       expect(results[0].hash, equals(1969199198234197));
       expect(results[0].distance, isNull);
 
@@ -363,9 +364,9 @@ void main() {
       results = await commands.georadiusbymember(
           key1, 'Madrid', 400.0, GeoUnit.kilometer,
           withCoord: true, withHash: true, withDist: true);
-      expect(results[0].member, 'Valencia');
-      expect(results[0].position.longitude, closeTo(-0.3545661, delta));
-      expect(results[0].position.latitude, closeTo(39.4561165, delta));
+      expect(results![0].member, 'Valencia');
+      expect(results[0].position!.longitude, closeTo(-0.3545661, delta));
+      expect(results[0].position!.latitude, closeTo(39.4561165, delta));
       expect(results[0].hash, equals(1969199198234197));
       expect(results[0].distance, closeTo(306.0561, delta));
 
@@ -384,14 +385,14 @@ void main() {
       results = await commands.georadiusbymember(
           key1, 'Madrid', 1000.0, GeoUnit.kilometer,
           order: GeoOrder.ascending);
-      expect(results[0].member, 'Madrid');
+      expect(results![0].member, 'Madrid');
       expect(results[1].member, 'Valencia');
       expect(results[2].member, 'Málaga');
 
       results = await commands.georadiusbymember(
           key1, 'Madrid', 1000.0, GeoUnit.kilometer,
           order: GeoOrder.descending);
-      expect(results[0].member, 'Málaga');
+      expect(results![0].member, 'Málaga');
       expect(results[1].member, 'Valencia');
       expect(results[2].member, 'Madrid');
 
@@ -509,22 +510,23 @@ void main() {
 
       group('GeoPosition', () {
         test('toString', () {
-          const value = GeoPosition(null, null);
+          const value = GeoPosition(0.0, 0.0);
           expect(value.toString(), startsWith('GeoPosition:'));
         });
       });
 
       group('GeoItem', () {
         test('toString', () {
-          const value = GeoItem<String>(null, null);
-          expect(value.toString(), startsWith('GeoItem<String>:'));
+          const position = GeoPosition(0.0, 0.0);
+          const value = GeoItem<String?>(position, null);
+          expect(value.toString(), startsWith('GeoItem<String?>:'));
         });
       });
 
       group('GeoradiusResult', () {
         test('toString', () {
-          const value = GeoradiusResult<String>(null);
-          expect(value.toString(), startsWith('GeoradiusResult<String>:'));
+          const value = GeoradiusResult<String?>(null);
+          expect(value.toString(), startsWith('GeoradiusResult<String?>:'));
         });
       });
     });
