@@ -3,13 +3,14 @@
 
 import 'package:test/test.dart';
 
+// ignore: directives_ordering
 import 'package:dartis/dartis.dart';
 
 import '../util.dart' show uuid;
 
 void main() {
-  Client client;
-  Commands<String, String> commands;
+  late Client client;
+  late Commands<String, String> commands;
 
   setUp(() async {
     client = await Client.connect('redis://localhost:6379');
@@ -105,7 +106,7 @@ void main() {
       await commands.set(key, 'abc');
 
       // Search.
-      expect(await commands.keys('$key'), equals([key]));
+      expect(await commands.keys(key), equals([key]));
       expect(await commands.keys('$key*'), isNotEmpty);
       expect(await commands.keys('${key}_'), isEmpty);
     });
@@ -299,7 +300,7 @@ void main() {
       await commands.set(key1, 'abc');
 
       // Dump and restore.
-      final value = await commands.dump(key1);
+      final value = (await commands.dump(key1))!;
 
       final key2 = uuid();
       await commands.restore(key2, 0, value);
