@@ -3,13 +3,14 @@
 
 import 'package:test/test.dart';
 
+// ignore: directives_ordering
 import 'package:dartis/dartis.dart';
 
 import '../util.dart' show uuid;
 
 void main() {
-  Client client;
-  ListCommands<String, String> commands;
+  late Client client;
+  late ListCommands<String?, String?> commands;
 
   setUp(() async {
     client = await Client.connect('redis://localhost:6379');
@@ -30,26 +31,26 @@ void main() {
 
       // Pop one element.
       var result = await commands.blpop(key: key1);
-      expect(result.key, equals(key1));
+      expect(result!.key, equals(key1));
       expect(result.value, equals('a'));
       result = await commands.blpop(key: key1);
-      expect(result.key, equals(key1));
+      expect(result!.key, equals(key1));
       expect(result.value, equals('b'));
 
       // Pop some elements.
       result = await commands.blpop(keys: [key1, key2]);
-      expect(result.key, equals(key1));
+      expect(result!.key, equals(key1));
       expect(result.value, equals('c'));
       result = await commands.blpop(keys: [key1, key2]);
-      expect(result.key, equals(key2));
+      expect(result!.key, equals(key2));
       expect(result.value, equals('d'));
       result = await commands.blpop(keys: [key1, key2]);
-      expect(result.key, equals(key2));
+      expect(result!.key, equals(key2));
       expect(result.value, equals('e'));
 
       // Pop blocking.
       result = await commands.blpop(keys: [key1, key2], timeout: 1);
-      expect(result.key, equals(key2));
+      expect(result!.key, equals(key2));
       expect(result.value, equals('f'));
 
       // Try to pop from an empty or non existing list.
@@ -68,26 +69,26 @@ void main() {
 
       // Pop one element.
       var result = await commands.brpop(key: key1);
-      expect(result.key, equals(key1));
+      expect(result!.key, equals(key1));
       expect(result.value, equals('c'));
       result = await commands.brpop(key: key1);
-      expect(result.key, equals(key1));
+      expect(result!.key, equals(key1));
       expect(result.value, equals('b'));
 
       // Pop some elements.
       result = await commands.brpop(keys: [key1, key2]);
-      expect(result.key, equals(key1));
+      expect(result!.key, equals(key1));
       expect(result.value, equals('a'));
       result = await commands.brpop(keys: [key1, key2]);
-      expect(result.key, equals(key2));
+      expect(result!.key, equals(key2));
       expect(result.value, equals('f'));
       result = await commands.brpop(keys: [key1, key2]);
-      expect(result.key, equals(key2));
+      expect(result!.key, equals(key2));
       expect(result.value, equals('e'));
 
       // Pop blocking.
       result = await commands.brpop(keys: [key1, key2], timeout: 1);
-      expect(result.key, equals(key2));
+      expect(result!.key, equals(key2));
       expect(result.value, equals('d'));
 
       // Try to pop from an empty or non existing list.
@@ -335,9 +336,9 @@ void main() {
 
       group('ListPopResult', () {
         test('toString', () {
-          const value = ListPopResult<String, String>(null, null);
+          const value = ListPopResult<String?, String?>(null, null);
           expect(
-              value.toString(), startsWith('ListPopResult<String, String>:'));
+              value.toString(), startsWith('ListPopResult<String?, String?>:'));
         });
       });
     });
