@@ -3,13 +3,14 @@
 
 import 'package:test/test.dart';
 
+// ignore: directives_ordering
 import 'package:dartis/dartis.dart';
 
 import '../util.dart' show uuid;
 
 void main() {
-  Client client;
-  PubSubCommands<String, String> commands;
+  late Client client;
+  late PubSubCommands<String?, String> commands;
 
   setUp(() async {
     client = await Client.connect('redis://localhost:6379');
@@ -57,8 +58,8 @@ void main() {
     group('support', () {
       group('PubsubResult', () {
         test('toString', () {
-          const value = PubsubResult<String>(null, null);
-          expect(value.toString(), startsWith('PubsubResult<String>:'));
+          const value = PubsubResult<String?>(null, null);
+          expect(value.toString(), startsWith('PubsubResult<String?>:'));
         });
       });
     });
