@@ -3,13 +3,14 @@
 
 import 'package:test/test.dart';
 
+// ignore: directives_ordering
 import 'package:dartis/dartis.dart';
 
 import '../util.dart' show uuid;
 
 void main() {
-  Client client;
-  SetCommands<String, String> commands;
+  late Client client;
+  late SetCommands<String, String> commands;
 
   setUp(() async {
     client = await Client.connect('redis://localhost:6379');
