@@ -75,7 +75,7 @@ class _TerminalDispatcher extends DispatcherBase {
   }
 
   @override
-  void onError(Object error, [StackTrace stackTrace]) {
+  void onError(Object error, [StackTrace? stackTrace]) {
     _controller.addError(error, stackTrace);
   }
 
