@@ -8,7 +8,7 @@ import '../protocol.dart';
 /// A transaction.
 class Transaction {
   /// Queued commands.
-  final List<Command<Object>> _queued = [];
+  final List<Command<Object?>> _queued = [];
 
   /// Whether this transaction is in progress.
   bool _inProgress = false;
@@ -17,7 +17,7 @@ class Transaction {
   bool get inProgress => _inProgress;
 
   /// Starts this transaction.
-  void begin(Command<Object> command) {
+  void begin(Command<Object?> command) {
     _inProgress = command is MultiCommand;
   }
 
@@ -28,7 +28,7 @@ class Transaction {
   }
 
   /// Completes a [command] with a [reply] in the context of this transaction.
-  void onReply(Command<Object> command, Reply reply, RedisCodec codec) {
+  void onReply(Command<Object?> command, Reply reply, RedisCodec codec) {
     assert(_inProgress);
     assert(command is! MultiCommand);
 
@@ -44,7 +44,7 @@ class Transaction {
   /// Completes a [command] with an error [reply] in the context of
   /// this transaction.
   void onErrorReply(
-      Command<Object> command, ErrorReply reply, RedisCodec codec) {
+      Command<Object?> command, ErrorReply reply, RedisCodec codec) {
     assert(_inProgress);
 
     // Failed EXEC command?
@@ -56,7 +56,7 @@ class Transaction {
   }
 
   /// Completes all commands in the transaction with [error].
-  void onError(Object error, StackTrace stackTrace) {
+  void onError(Object error, StackTrace? stackTrace) {
     assert(_inProgress);
 
     for (final command in _queued) {
@@ -67,18 +67,17 @@ class Transaction {
   }
 
   /// Completes all the queued commands.
-  void _exec(Command<Object> command, Reply reply, RedisCodec codec) {
+  void _exec(Command command, Reply reply, RedisCodec codec) {
     // Redis server replies a null value when some watched keys are modified.
     if (reply.value == null) {
       _discard(command, reply, codec);
     } else {
-      // ignore: avoid_as
       _dequeue(command, reply as ArrayReply, codec);
     }
   }
 
   /// Completes all the queued commands with an error.
-  void _discard(Command<Object> command, Reply reply, RedisCodec codec) {
+  void _discard(Command command, Reply reply, RedisCodec codec) {
     final error = ErrorReply('Transaction discarded.'.codeUnits);
 
     for (final command in _queued) {
@@ -100,7 +99,7 @@ class Transaction {
   }
 
   /// Enqueues a [command].
-  void _enqueue(Command<Object> command, Reply reply, RedisCodec codec) {
+  void _enqueue(Command<Object?> command, Reply reply, RedisCodec codec) {
     if (reply is! StringReply) {
       throw RedisException(
           'Expected "StringReply", but "${reply.runtimeType}" found instead.');
@@ -117,7 +116,7 @@ class Transaction {
   }
 
   /// Completes the queued commands with the array of replies in [reply].
-  void _dequeue(Command<Object> command, ArrayReply reply, RedisCodec codec) {
+  void _dequeue(Command command, ArrayReply reply, RedisCodec codec) {
     final array = reply.array;
 
     if (array.length != _queued.length) {
