@@ -11,17 +11,17 @@ abstract class HyperLogLogCommands<K, V> {
   /// altered, `0` otherwise.
   ///
   /// See https://redis.io/commands/pfadd
-  Future<int> pfadd(K key, {V element, Iterable<V> elements = const []});
+  Future<int?> pfadd(K key, {V? element, Iterable<V> elements = const []});
 
   /// Returns the approximated cardinality of the set(s) observed by the
   /// HyperLogLog at key(s).
   ///
   /// See https://redis.io/commands/pfcount
-  Future<int> pfcount({K key, Iterable<K> keys = const []});
+  Future<int?> pfcount({K? key, Iterable<K> keys = const []});
 
   /// Merges multiple HyperLogLog values into a single one.
   ///
   /// See https://redis.io/commands/pfmerge
   Future<void> pfmerge(K destkey,
-      {K sourcekey, Iterable<K> sourcekeys = const []});
+      {K? sourcekey, Iterable<K> sourcekeys = const []});
 }
