@@ -9,7 +9,7 @@ import 'reply.dart';
 
 /// A converter that converts an instance of type [S] into an instance
 /// of type [T].
-abstract class Converter<S, T> {
+abstract class Converter<S extends Object, T extends Object> {
   /// The source type [S] of this converter.
   Type get sourceType => S;
 
@@ -20,18 +20,31 @@ abstract class Converter<S, T> {
   T convert(S value, RedisCodec codec);
 
   /// Checks if this can converts a [value] into an instance of type [U].
-  bool supports<U>(Object value) => value is S && U == targetType;
+  bool supports<U>(Object value) =>
+      value is S && U == targetType ||
+      U.toString() == '${targetType.toString()}?';
 }
 
 /// A converter that converts an instance of type [S] into a list of bytes.
 ///
 /// Extend from this class for building your own encoders.
-abstract class Encoder<S> extends Converter<S, List<int>> {}
+abstract class Encoder<S extends Object> extends Converter<S, List<int>> {}
 
 /// A converter that converts a server reply into an instance of type [T].
 ///
 /// Extend from this class for building your own decoders.
-abstract class Decoder<S extends Reply, T> extends Converter<S, T> {}
+abstract class Decoder<S extends Reply, T extends Object>
+    extends Converter<S, T> {}
+
+// A convert that converts a server reply into a list of instance of type [T].
+///
+/// Extend from this class for building your own array decoders.
+abstract class ArrayDecoder<S extends Reply, T> extends Converter<S, List<T>> {
+  /// Checks if this can converts a [value] into an instance of type [List<U>].
+  @override
+  bool supports<U>(Object? value) =>
+      value != null && value is S && U.toString().startsWith('List<$T>');
+}
 
 /// A generic converter that encodes instances of any type to lists of bytes
 /// and decodes lists of server replies to instances of any type.
@@ -47,7 +60,8 @@ class RedisCodec {
 
   /// Registers a new [encoder] or a new [decoder], or both.
   void register(
-      {Converter<Object, Object> encoder, Converter<Object, Object> decoder}) {
+      {Converter<Object?, Object?>? encoder,
+      Converter<Object?, Object?>? decoder}) {
     assert(encoder != null || decoder != null);
 
     if (encoder != null) {
@@ -67,13 +81,13 @@ class RedisCodec {
 
 /// A converter that converts instances of multiple types.
 abstract class _MultiConverter {
-  final List<Converter<Object, Object>> _converters = [];
+  final List<Converter<Object?, Object?>> _converters = [];
 
   /// Registers a new [converter].
   ///
   /// If a converter for the given types already exists the new one replaces
   /// the existing one.
-  void register(Converter<Object, Object> converter) {
+  void register(Converter<Object?, Object?> converter) {
     log.fine(() => 'Registering converter: $converter.');
 
     _converters
@@ -87,7 +101,6 @@ abstract class _MultiConverter {
   T convert<T>(Object value, RedisCodec codec) {
     for (final converter in _converters) {
       if (converter.supports<T>(value)) {
-        // ignore: avoid_as
         return converter.convert(value, codec) as T;
       }
     }
@@ -119,8 +132,11 @@ class _Decoder extends _MultiConverter {
     register(_DoubleReplyDecoder());
     register(_ArrayReplyDecoder<List<int>>());
     register(_ArrayReplyDecoder<String>());
+    register(_ArrayReplyDecoder<String?>());
     register(_ArrayReplyDecoder<int>());
+    register(_ArrayReplyDecoder<int?>());
     register(_ArrayReplyDecoder<double>());
+    register(_ArrayReplyDecoder<double?>());
   }
 }
 
@@ -133,13 +149,13 @@ class _RawEncoder extends Encoder<List<int>> {
 /// An encoder that converts a [String] into a list of bytes.
 class _StringEncoder extends Encoder<String> {
   @override
-  List<int> convert(String value, RedisCodec codec) => utf8.encode(value);
+  List<int> convert(String? value, RedisCodec codec) => utf8.encode(value!);
 }
 
 /// An encoder that converts an [int] into a list of bytes.
 class _IntEncoder extends Encoder<int> {
   @override
-  List<int> convert(int value, RedisCodec codec) => value.toString().codeUnits;
+  List<int> convert(int? value, RedisCodec codec) => value.toString().codeUnits;
 }
 
 /// An encoder that converts a [double] into a list of bytes.
@@ -149,8 +165,8 @@ class _DoubleEncoder extends Encoder<double> {
   static const List<int> _negativeInfinity = <int>[45, 105, 110, 102]; // -inf
 
   @override
-  List<int> convert(double value, RedisCodec codec) {
-    if (value.isFinite) {
+  List<int> convert(double? value, RedisCodec codec) {
+    if (value!.isFinite) {
       return value.toString().codeUnits;
     }
     if (value == double.infinity) {
@@ -168,32 +184,27 @@ class _DoubleEncoder extends Encoder<double> {
 /// A decoder that returns the raw value of a server reply.
 class _RawReplyDecoder extends Decoder<SingleReply, List<int>> {
   @override
-  List<int> convert(SingleReply value, RedisCodec codec) => value.bytes;
+  List<int> convert(SingleReply? value, RedisCodec codec) => value!.bytes;
 }
 
 /// A decoder that converts a server reply into an [String].
 class _StringReplyDecoder extends Decoder<SingleReply, String> {
   @override
   String convert(SingleReply value, RedisCodec codec) =>
-      value.bytes == null ? null : utf8.decode(value.bytes);
+      utf8.decode(value.bytes);
 }
 
 /// A decoder that converts a server reply into an [int].
 class _IntReplyDecoder extends Decoder<SingleReply, int> {
   @override
   int convert(SingleReply value, RedisCodec codec) =>
-      value.bytes == null ? null : int.parse(String.fromCharCodes(value.bytes));
+      int.parse(String.fromCharCodes(value.bytes));
 }
 
 /// A decoder that converts a server reply into a [double].
 class _DoubleReplyDecoder extends Decoder<SingleReply, double> {
   @override
   double convert(SingleReply value, RedisCodec codec) {
-    if (value.bytes == null) {
-      // ignore: avoid_returning_null
-      return null;
-    }
-
     final number = String.fromCharCodes(value.bytes);
 
     if (number == r'+inf') {
@@ -209,9 +220,9 @@ class _DoubleReplyDecoder extends Decoder<SingleReply, double> {
 
 /// A decoder that converts an array of server replies into a list of
 /// instances of type [T].
-class _ArrayReplyDecoder<T> extends Decoder<ArrayReply, List<T>> {
+class _ArrayReplyDecoder<T> extends ArrayDecoder<ArrayReply, T> {
   @override
-  List<T> convert(ArrayReply value, RedisCodec codec) => value.array == null
-      ? null
-      : value.array.map((reply) => codec.decode<T>(reply)).toList();
+  List<T> convert(ArrayReply value, RedisCodec codec) => value.array
+      .map((reply) => reply is NullReply ? null as T : codec.decode<T>(reply))
+      .toList();
 }
