@@ -1,8 +1,8 @@
 // Copyright (c) 2018, Juan Mellado. All rights reserved. Use of this source
 // is governed by a MIT-style license that can be found in the LICENSE file.
 
-import 'dart:io' show BytesBuilder;
-
+import 'dart:io';
+import 'dart:typed_data' show BytesBuilder;
 import 'codec.dart';
 import 'token.dart';
 
