@@ -3,6 +3,7 @@
 
 import 'package:test/test.dart';
 
+// ignore: directives_ordering
 import 'package:dartis/dartis.dart';
 
 void main() {
@@ -155,7 +156,7 @@ void main() {
 
       // EXEC
       final exec = ExecCommand(<Object>['EXEC']);
-      const reply = BulkReply(null);
+      const reply = NullReply();
       transaction.onReply(exec, reply, codec);
 
       expect(set.future, throwRedisException);
