@@ -3,6 +3,7 @@
 
 import 'package:test/test.dart';
 
+// ignore: directives_ordering
 import 'package:dartis/dartis.dart';
 
 void main() {
@@ -56,7 +57,7 @@ void main() {
       });
 
       test('toString', () {
-        const reply = BulkReply(null);
+        const reply = BulkReply([]);
 
         expect(reply.toString(), startsWith('BulkReply:'));
       });
@@ -87,7 +88,7 @@ void main() {
       });
 
       test('toString', () {
-        const reply = ArrayReply(null);
+        const reply = ArrayReply([]);
 
         expect(reply.toString(), startsWith('ArrayReply:'));
       });
